refactor: update import paths for utility and article types
