+++ src/components/atoms/GithubArticleButton.tsx
@@ -1,5 +1,5 @@
 import { FC } from 'react';
-import { Maybe, Optional } from '../../../types/utility';
+import { Maybe, Optional } from '../../types/utility';
 import { metaData } from '../../constants/metaData';
 
 type Props = {
