+++ src/pages/blog/[id].tsx
@@ -2,8 +2,8 @@ import { JSX } from 'react';
 import { GetStaticPaths, GetStaticProps } from 'next';
 import { getArticleMetaData } from '../../libraries/articles';
 import { PageTemplate } from '../../templates/PageTemplate';
-import { ArticleMetaData } from '../../../types/article';
-import { Optional } from '../../../types/utility';
+import { ArticleMetaData } from '../../types/article';
+import { Optional } from '../../types/utility';
 import { XShareButton } from '../../components/atoms/XShareButton';
 import { GithubArticleButton } from '../../components/atoms/GithubArticleButton';
 import { MdTemplate } from '../../templates/MdTemplate';
