+++ src/types/@mapbox__rehype-prism.d.ts
